requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.6.15=pyhd8ed1ab_1
cffi=1.15.1=py310h74dc2b5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
cryptography=37.0.2=py310h597c629_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.37.1=pypi_0
freetype=2.10.4=h0708190_1
geatpy=2.7.0=pypi_0
giflib=5.2.1=h36c2ea0_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jbig=2.1=h7f98852_2003
jpeg=9e=h166bdaf_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
lerc=2.2.1=h9c3ff4c_0
libdeflate=1.7=h7f98852_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_1
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.5.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310ha2c4b55_0
mkl_fft=1.3.1=py310h2b4bcf5_1
mkl_random=1.2.2=py310h00e6091_0
mmcls=0.23.2=pypi_0
mmcv-full=1.6.0=pypi_0
mmdet=2.25.1=pypi_0
mmsegmentation=0.28.0=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
numpy=1.23.1=py310h1794996_0
numpy-base=1.23.1=py310hcba007f_0
opencv-python-headless=4.6.0.66=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1q=h7f8727e_0
ordered-set=4.1.0=pypi_0
packaging=21.3=pypi_0
pillow=9.2.0=py310hace64e9_1
pip=22.1.2=py310h06a4308_0
prettytable=3.4.1=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pypi_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_1
setuptools=63.4.1=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.2=h5082296_0
terminaltables=3.1.10=pypi_0
tk=8.6.12=h1ccaba5_0
torchvision=0.13.1=py310_cu116
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022a=hda174b7_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yapf=0.32.0=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.0=ha95c52a_0
